refactor(site-docs): extract scattered pins and features grid into components
Move SCATTERED_PINS data + rendering into ScatteredPins component and
FEATURES data + grid section into FeaturesGrid component. Page file
is now a thin composition layer at 81 lines.
